Optimise benchmarks
