+++ benchmarks/src/jmh/scala/info/folone/scala/poi/benchmarks/AsyncPoiBenchmarks.scala
@@ -12,11 +12,11 @@ import scala.concurrent.duration._
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
 @Fork(1)
-@Warmup(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
-@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Warmup(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
 class AsyncPoiBenchmarks {
 
-  @Param(Array("100", "1000"))
+  @Param(Array("100"))
   var dataSize: Int = _
 
   var workbook: Workbook = _
